chore: TokenStorage 파일 분리
